5830 FIX Fixed DST shift correction for downtimes, causing CMC to use 100% CPU
When a downtime was configured to recur every hour and a 1h DST shift
happened, the CMC would go into an infinite loop with 100% CPU load and no
monitoring at all.  If the downtime in question was set in an ad hoc fashion
via the GUI (not via rules), the only way to work around this issue was to
remove the state file, losing all downtimes, acknowledgements and comments.

Change-Id: Iebca5c808b39947682da235b08dffb73a883299b