5946 FIX lnx_if: Exclude veth* network interfaces on docker nodes
The veth* network interfaces created for docker containers are now
excluded by the linux agent in all cases. The interface names have no
direct match with the docker container name or ID. They seem to have
some kind of random nature.

These container specific interfaces are not relevant to be monitored
on the node. We are monitoring the docker network interfaces in the
container.

Change-Id: Icecc44425b30cd1ae4e74448f7f71f86ab7b065e