Rename WinAPI wrapper base class to WinApiInterface
The naming WinApiAdaptor misleadingly reminded of Adaptor Pattern,
which this is implementation is (at least not yet) following. The
current implementation rather reminds of the Facade Pattern. This may
change in the future, though, as there would be a need for adapting the
data types passed through the wrapper class in order to cut direct
dependencies to Microsoft-specific data types. Change the naming of the
abstract base wrapper class to describe its role as an "interface" with
exclusively pure virtual member functions.