5947 df: Exclude docker local storage mounts on docker nodes
The df check is now excluding all filesystems found below
<tt>/var/lib/docker</tt>, which is the default location for
the docker container local storage.

Depending on the used storage engine docker creates overlay
filesystems and mounts below this hierarchy for the started
containers.

The filesystems are not interesting for our monitoring. They
will be monitored from the container context.

Change-Id: I25deab75bf7e9cc1a73b34999ce046f0bb06092a