5833 FIX Fixed handling of 'acknowledged' parameter of the EC's UPDATE command
Previously, the 'acknowledged' parameter of the EC's UPDATE command was
processed after the event was already (partially) updated.  When the
'acknowledged' parameter was missing or not an integer, this behavior was
confusing and could e.g. lead to missing history entries.

Now all parameters are checked before the event is modified, so no partial
updates happen.  Furthermore, the 'acknowledged' parameter is optional now,
restoring the previous accidental behavior in a clean way.

Change-Id: I29e44d2c54b6b01ceaad8415c5012b5206cbc50d