Make _what an instance attribute for custom host/user attributes
Previously, _what was a class attribute defined in subclasses, but was
used already in the base classes. Now, _what is made an instance
attribute defined in the base class itself. This makes the use of _what
more explicit. This is sufficient to make pylint happy again.

In the long run _what has to be moved to the subclasses, because
base classes should not now implementation details of child classes.

Change-Id: I19edbe611a115aab26a28009e337981636b8e405