5975 FIX Agent output / SNMP walk download: Now processed asynchronously
The download of the agent output or SNMP walks, that can be triggered from the
hosts or Check_MK service context menus could run into time out situations with
bad error messages like "Gateway timeout" in case the destination devices need
more than ~110 seconds to respond.

We have now changed the process of fetching the data to be executed as
background job. The GUI is displaying the progress during processing and
presenting a download icon on the progress page to finally download the created
file.

Change-Id: Ia3fc705454d2b5409bc852421202b8df5bf104ad