5275 FIX cmk-update-agent: Fix run-as-plugin mode (Regression since 1.5.0b1)
Due to an erroneous call to a non-existant method, the Agent Updater crashes when run as plugin.

As a result, automatic updates would not work with an Agent Updater 1.5.0b1. Manual updates are still working, so it is possible to replace the erroneous Agent Updater with a fixed one by doing a manual Update to a newly baked agent.

Change-Id: I941d868bca1df06e5a097e84c34c180fa09b3699