5983 FIX docker_node_images: Fixed mixups between images of same repository and different tags
The number of containers was not counted per image because the tag information
was not handled correctly in all cases.

Some containers don't explicitly state "repository:tag" but only "repository"
as their image, depending on the "docker run" call. Check_MK is now falling
back to "latest" tag in this case.

Change-Id: I2fc33ced414d040eaf9dde683d1478cdc288d653