5913 FIX webapi: Fixed handling of the option mode in service discovery
The previous implementation of the option mode in service_discovery did not
allow a handling in the same way as in other webapi requests like including it
into the request part. This is now fixed. The old behaviour is still supported.

old only supported this syntax:
curl http://localhost//check_mk/webapi.py?action=discover_services

Change-Id: If2c5d149332aca75aaebf8cde3f04e1a86230f16