5985 FIX htpasswd command is now available in all sites
The "omd create" success message tells users about using the htpasswd
command to change the password of the default GUI user. On SLES platforms
the command is available under the name "htpasswd2". We provide a symlink
to the platform specific name to make the command available for all site
users now.

Change-Id: I4f75d6a81c4417d92b3c11c91edf64c6cd1ba42a