5923 FIX Windows: section MRPE was incorrectly formatted
The output of section MRPE was incorrectly formatted. A null character
at the end of the output string caused the output to be cut abruptly
and missing a newline character before the following section header.
Broken by commit 6503810d95eddff68cd9c7faabe8d64c15c5cc02 that was
included in releases 1.5.0i3 and 1.5.0b1.