5923 FIX Windows: section MRPE was incorrectly formatted
The output of section MRPE was incorrectly formatted. A null character
at the end of the output string caused the output to be cut abruptly
and missing a newline character before the following section header.
Broken by commit be3a927001ff04ab419a45b392cf59035b7e1602 that was
included in releases 1.5.0i3 and 1.5.0b1.