6034 FIX Fixed adding combined graphs to reports
When clicking on the context menu of a combined graph and adding it to a report,
the element that is added to the report is not created correctly. When trying
to edit the element later an exception "TypeError (unhashable type: 'dict')" is
shown.

Delete the element after the update and re-add it to the report should fix the
issue.

Change-Id: I88bc3c6680e2f8b2a7486ff34b177dcdf84285b0