Don't import names directly, use their hierarchical names.
Explicit is better than implicit, and the previous naming was confusing
(even if it was shorter).  Actually the naming was so confusing that it
contained a bug...  :-) This has been fixed on the way, a werk is coming in
a separate commit.

With some global names now free, rename some local variables back to
something more readable.

Change-Id: Iab5ba58aee8ee1007bf01967bb2aeb71da348cea