6046 Improved execution time of e.g. service discovery
The overall execution time of the service discovery has been reduced by
introducing some precompiling of the check plugins and their includes
before loading them. The mechanism is similar to the compilation of
python modules from (.py to .pyc files). This mechanism reduces the
load time of the plugins by 50%.

Change-Id: I97bc4f72db4fa782d183fd936863d3bb87c25c90