6053 FIX Fixed sync after user notification rule changes
When a user changes his personal notification rules in a distributed
environment, the user profile synchronization was only performed for the
configured "login sites". These are the sites where user logins are allowed (by
the site configuration). In case of notification setting changes the notification
rules need to be synchronized to all sites instead.

Change-Id: I6d9f4b9ee0cecba116d16fc917973c04ca610ff5