5810 FIX CMC Graphs: fixed incorrect fallback to pnp4nagios graphs in case some sites are disabled
The detection routine, which determines if CMC-graphs are possible could fail when a remote site
was down or unreachable. Furthermore, it didn't matter if the site in question had the actual
rrd data for the graph, so it could have been any site.

As fallback, pnp4nagios graphs were used - which may lead to other display errors.

Change-Id: Icc194b2edb92983991d560813d8725ef4de25346