CMK-517: Add Werk 5277
5277 cmk-update-agent: Add support for proxy servers

The agent updater is now able to connect to the Check_MK server via proxy.

With this werk, a new checkbox "Use proxy server" is added to the Agent Updater rule configuration. You can now either configure a HTTP, SOCKS4 or SOCKS5 proxy (with optional credentials) or choose to follow the environment variables "HTTP_PROXY" and "HTTPS_PROXY" that may be set locally on the hosts.

Please note that the configuration default is still set to ignore local proxy settings.

Change-Id: I795942a25a5435956589eff21de6f4e7bf28e17f