6099 FIX Only reset default version when current default version is uninstalled
Previously all RPM packages that were uninstalled were resetting the current
default version to the result of "ls -A /omd/versions 2>/dev/null| sort -n |
tail -n 1". The uninstall logic of NEW packages will now first check whether or
not the version currently being removed is the default version and only reset
the default version in case it is needed.

Change-Id: Iee25acb6bc6352256a60ffa10f9c8654a9956c2d