6101 FIX Fixed partial uninstall of Check_MK 1.5 versions (leaving sitecustomize.pyc)
The python delivered with Check_MK was not precompiling the lib/python2.7/sitecustomize.pyc
during packaging. It was then created during runtime on the target system. Since this file
is not under control of the RPM/DEB packages the file was left on the filesystem after
uninstall. The version was still visible after removal in "omd version" output until one
removed the /omd/versions/[version] directory.

Change-Id: Ie7eb6e4b68589f598dfd7c654f2de87c7cc8ff15