6107 SEC Agent bakery signing key passphrases could be visible in access logs
When you are using the agent bakery for creating and distributing your
monitoring agents it is likely that this change is relevant for you.

In some parts of the GUI, when working with agent signing keys, it could happen
that the signing key secret you enter could be written to the apache access log
of your Check_MK server. This means that it may be visible to all local system
users (e.g. users with access to the command line).

This affects the access log of the system apache (normally located at
/var/log/apache2) and the access log (/omd/sites/[site]/var/log/apache/*) of
the sites apache (master site in distributed setups).

You may want to scan the log files for the string "key_p_passphrase" to check
whether or not you are affected. It can be done e.g. with:

zgrep key_p_passphrase /var/log/apache2/access* /omd/sites/*/var/log/apache/access*

In case you find something, you should clean it up. Delete the file or remove
the secrets from that file.

Even if it's unlikely that your key has been compromised, it is recommended to
stop using this signing key. Create a new key and proceed with this one.

Change-Id: I2d7ab495596a24baeabf601c9b327b1757c1d106