6151 FIX Net-SNMP: Add some patch trying to fix SNMPv3 communication in some cases
When using SNMPv3 communication issues with some devices may occur. From the user point of
view timeouts are reported by Check_MK originating from Net-SNMP not receiving a valid response.

This kind of issue is related to a broken negotiation (related to the maximum message size) between
the monitoring and the monitored device.

Change-Id: I310a0f59867130183cc5184bc2195f109c306776