Fix dependencies of core-related modules, part 1.
We are doing two things wrong here:

First of all, we abuse modules as "normal" data structures, mixing up the
concepts of a class and its instances into a single concept (the module).
What we really want is a class definition and the explicit usage of its
instances.

This is already bad, but things get worse: In this pseudo-OO world, the
"superclass" knows its "subclasses", which violates two fundamental design
principles:

 * the Liskov substitution principle
   (see https://en.wikipedia.org/wiki/Liskov_substitution_principle)

 * the open/closed principle
   (see https://en.wikipedia.org/wiki/Open/closed_principle)

In our concrete example, the more general core_config module knows about the
more concrete core_nagios and core_cmc modules.  Apart from violating the
two principles above, this is the root cause of several import cycles.

As a first tiny step, this commit unified the interface for precompilation.
More to come...

Change-Id: I5f39b88afff07c29643b5cd3926874c7b2a8bc08