Fix dependencies of core-related modules, part 5.
The hook for config generation is now passed down the call hierarchy.  This
removes one part of the wrong dependency direction, the other part
(precompilation hook) will follow later.

Note that we introduce a tiny bit of duplication in AutomationRestart, but
this is already a copy-n-paste hell, so this should be fixed later, anyway.

The "threading through" part is a direct consequence of the current non-OO
design and will vanish when we have an object representing the configured
core (Nagios or CMC).

Change-Id: I1230efa6c05c988ea6882bc2192b2d48832d61a8