6190 FIX Win-agent: prevent unsigned integer overflow in process uptime
The process uptimes for Windows are calculated by subtracting the process
creation time from the current system time. Under certain circumstances,
setting up the system clock e. g. with daylight saving time has led to
some processes reporting a creation time with false offset and an unsigned
integer overflow through negative subtraction result. This has further led
to the crash of the ps check.

The unsigned integer overflow is now prevented by checking the result of the
subtraction and, in case of a negative value, logging it as an error and
setting the process uptime to the default value 1.

Change-Id: I4cbefc5e500880594be0f385d86cc3f1959ce683