Removed a piggyback-related import cycle.
Although this removes a cycle, pylint now shows 7 cycles: For some obscure
reason it finds 2 more cycles now, but those are not related to this commit.
We really need to investigate what's goning on here, but to do this we
should really, really disentangle pytest and pylint, the current state is
totally arcane and not comprehensible at all...  :-/

Back to the commit itself: To break the cycle, we thread through a
parameter, although this is a bit verbose and breaks the piggyback
"abstraction" a bit.  Nevertheless, breaking dependency cycles is *far* more
important than this tiny ugliness, and what we really want here is a class:
This would fix both problems at once, and probably many more...

Change-Id: I890f8d5666f5dd4551a2c52ae00fa7d05e8eb807