5824 FIX Agent Bakery: Fixed missing baked packages
During the agent bake process some previously generated symlinks went missing.
The baking itself went fine, but the symlinks pointing to the bake result were no longer available.
As a result, the agent bakery GUI showed some packages as "Not available".
This is a preliminary fix, until we find the root cause of this problem.

There is also a console woraround for this problem.
Agent baking will always work with the command <tt>cmk -v --bake-agents</tt>

Change-Id: I5fc7365e4c95d791d9f8b32fdb2ec58f058914b4