Removed the last import cycle in cmk_base.
This was done by merging cmk_base.{config,checks,rulesets} back into a
single module again.  The initial intention of splitting this rather big
module up was well-intentioned, but the split simply did not reflect
reality: Each of the three modules was heavily tied to the other two
modules, resulting in 3 import cycles between them.  So in effect, the split
was wishful thinking, the module effectly stayed a single entity, only the
source code was scattered around 3 different files.

Import cycles are an absolute architectural no-no, and in addition they
actually harm most kinds of tooling heavily.  So having a larger module
again is a very small price to pay...

Change-Id: If62d400466eecbee76f38135946daa1c7863ec1a