6224 FIX Prevent multiple mkeventd's from being started in some situations
In rare circumstances, a non-responsive mkeventd process would stick
around after an OMD restart, and the new one would not be aware of
the other one's existance due to duplicated cleanup of the PID file
in the Event Console's init script and the process itself respectively.

We remove the cleanup in the init script, as it is more coarse.

Change-Id: I7f50835dbcb4b52f0ebf3b1fd0e81d886fa1b1fb