6324 FIX Notifications: Fixed height of graph area with a larger legend
Graphs with a lot of metrics need more space for their legend. This larger
legend did not increase the size of the graph images. They reduced the size
of the graph are and finally destroyed the graph layout.

The image height is now increased with a larger number of metrics.

Change-Id: I44b06be464283e7a25a2c48ff2eac6a5e04764d5