Fixed and enabled keyword-arg-before-vararg warnings.
Note that there was actually no way to pass "enabled" as a keyword argument
and pass extra positional arguments as well.  So Profile.__init__'s
signature didn't make sense, but the fix was easy: Just remove positional
arguments. The same holds for ImageUpload.__init__.

CMKWebLogger.exception is a different story: Its signature differs from the
one of the superclass method, which is ugly and a design no-no.  Rather than
fixing all the call sites, we introduce some local Kung Fu...

Change-Id: Ic9a09d40491ef224073056ca8192d2c10cdb8669