CMK-563: Add Werk 5502
5502 FIX Prevent baked tar.gz agents from changing ownership of files on host

When extracting a tar.gz packaged agent from the agent bakery as root (as it is normallly done), the user and group ownerships of the extracted files would be set to the Check_MK site user, if such user is available on the target host. This would lead to a wrong ownership on some folders like "/etc", "/usr" and "/var".

In order to prevent this, the agent bakery now creates the tar.gz packages correctly by setting an ownership of "root".

Change-Id: I0337aaddf8bed83a8cc03c5908415fabbd0be40f