CMK-558: Add Werk 5503
5503 cmk-update-agent: Convenient logging on Linux

The Agent Updater now makes use of its logging capabilities more conveniently.
Only ERRORs and WARNINGs will be logged to syslog.

Additionally, a detailed log will be created at /var/lib/check_mk_agent/cmk-update-agent.log
when running as plugin. It will be rotated every 5MB with a maximum of 3 backups.
This improves our capabilities for an a-posteriori error analysis.

When running cmk-update-agent manually,
a logfile can be specified using the -l flag.

Change-Id: I0ab5390f3ab5f9d50dbc75ec45c752f5d6235973