6342 FIX Fixed degraded SNMP check performance when using SNMP sub checks
When checking a host with SNMP sub checks which normally all get the same data, the
data was fetched multiple times. One time for each subcheck used for monitoring a host.

This resulted in significantly increased check times and time outs. All 1.5 versions
are affected by this issue.

We experienced this issue with a CMCiii device which is normally monitored by a set of
sub checks (cmciii.*).

Change-Id: I83494499ac31fa354b463388068e6551aea36a54