Dropped the automatically available html variable
Previously the current html object was available to all GUI code via
the __builtin__ namespace while processing a request. This was tool
friendly and has been replaced with this mechanism. We have a proxy object
which forwards all requests to the html() object of the current request.

Each module that needs the html object for e.g. rendering html code
needs to use the import "from cmk.gui.globals import html" first.

Change-Id: Id590d42bc0a8a27defeb6bad2619c6f0dd00bdb8