6345 FIX Fixed invisible/lost notification rules when editing notification rules
When cloning notification rules or creating notification rules with identical
parameters only one of the rules was shown. The others could be hidden, but
still be in existing.

In other situations, when a user with limited permissions on notification plugins
edits the notification rules, it could happen that all notification rules that
use other plugins were deleted by accident.

These problems were introduced with werk #4167.

CMK-556

Change-Id: Icf56d0dca51d47e89a0c402819b73aaa39952704