6350 FIX Fixed broken checking of unrelated hosts when using manual checks
When using the Check_MK Raw Edition it could happen that configuring manual
checks for a host broke the check execution of unrelated hosts.

The configuration of the manual check of all hosts was loaded into the context
of such an unrelated host. Because a host only loads the check definitions of
the checks it needs during checking, those manual check configurations were
trying to access unavailable data. This lead to a crash and a "(null)" output
shown for the "Check_MK" service.

Change-Id: I7347801936d7e88309a09efc36756abb7f57a6ef