6127 FIX ucd.include: Astaro Firewall is Linux but should use hr_mem
Astaro firewalls are Linux but should use hr_mem. Otherwise caches/buffers are
included in used memory generating critical a state.

To switch to hr_mem on Astaro firewalls the services on the affected hosts have
to be rediscovered. After the rediscovery the ucd_mem service will become
undecided and the hr_mem services will be monitored. The undecided services
will only be shown for a short time and will automatically disappear when the
caches in Check_MK are updated. I.e. after the rediscovery now further actions
are necessary.

Change-Id: I0b953cf704c112468b947661bd3480587afaa8e9