Fixed background job standard FD handling
After each background job start these messages could be found in
the apache error log:

  [Tue Jul 24 20:06:01.490903 2018] [wsgi:error] [pid 16596] [client
  127.0.0.1:40244] close failed in file object destructor:
  [Tue Jul 24 20:06:01.490960 2018] [wsgi:error] [pid 16596] [client
  127.0.0.1:40244] IOError: [Errno 9] Bad file descriptor

This problem was caused by the handling of the open file descriptors
when starting background job.

During start all FDs except the first 3 were closed. The sys.stdout
object was replaced in python with a StringIO() wrapper which is
then used to collect the job results. But the FDs 0, 1 and 2 were
not closed.

To have a clean initial state we are now closing all FDs without
exception. We proceed without stdin and also replace stderr with a
stream wrapper that sends the written data to the var/log/web.log.

Change-Id: I3596a793ac37d406b0aeded15cb5f6df922cee0b